fix: allow missing types of rule bypasses
Leave it to the API to determine if the values for actor_types are valid
of not. These change over time and we would need to carry a list in code
to verify this. Do not require an actor_id on rule bypasses. Things like
DeployKey do not have an id.

Signed-off-by: Johannes Wienke <languitar@semipol.de>